# contracts/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: api-tracker
  version: "1.0"
  description: API-tracker REST API. Base URL — https://api.apitracker.ru
servers:
  - url: https://api.apitracker.ru
    description: Production
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: JWT access token (RS256, 15 min) or PAT (prefix `pat_`).
  schemas:
...
    ProjectPermissions:
      type: object
      properties:
        edit_title: { type: boolean }
        edit_description: { type: boolean }
        edit_tags: { type: boolean }
        edit_blockers: { type: boolean }
        edit_assignee: { type: boolean }
        edit_status: { type: boolean }
        share: { type: boolean }
        delete_task: { type: boolean }
        rename_project: { type: boolean }
        manage_members: { type: boolean }
        manage_automations: { type: boolean }
        manage_attachments: { type: boolean }
        delete_project: { type: boolean }
        import_tasks: { type: boolean }
...
    ProjectInvitation:
      type: object
      properties:
        id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        invitee_user_id: { type: string, format: uuid, nullable: true }
        invitee_team_id: { type: string, format: uuid, nullable: true }
        invited_by: { type: string, format: uuid }
        permissions: { $ref: "#/components/schemas/ProjectPermissions" }
        status: { type: string }
        created_at: { type: string, format: date-time }
      required: [id, project_id, invited_by, permissions, status, created_at]
    TeamInvitation:
      type: object
      properties:
        id: { type: string, format: uuid }
        team_id: { type: string, format: uuid }
        invitee_user_id: { type: string, format: uuid }
        invited_by: { type: string, format: uuid }
        status: { type: string }
        created_at: { type: string, format: date-time }
      required: [id, team_id, invitee_user_id, invited_by, status, created_at]
...
security:
  - BearerAuth: []
paths:
...
  /projects/{id}/invitations:
    post:
      summary: Create project invitation
      operationId: createProjectInvitation
      tags: [Invitations]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                invitee_user_id: { type: string, format: uuid }
                invitee_team_id: { type: string, format: uuid }
                permissions: { $ref: "#/components/schemas/ProjectPermissions" }
              required: [permissions]
      responses:
        "201":
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectInvitation"

    get:
      summary: List project invitations
      operationId: listProjectInvitations
      tags: [Invitations]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: List of project invitations
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ProjectInvitation"
                required: [items]
  /projects/{id}/invitations/{invitation_id}:
    delete:
      summary: Delete project invitation
      operationId: deleteProjectInvitation
      tags: [Invitations]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: invitation_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Invitation deleted
  /invitations/projects/{invitation_id}/accept:
    post:
      summary: Accept project invitation
      operationId: acceptProjectInvitation
      tags: [Invitations]
      parameters:
        - name: invitation_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Invitation accepted
  /invitations/projects/{invitation_id}/decline:
    post:
      summary: Decline project invitation
      operationId: declineProjectInvitation
      tags: [Invitations]
      parameters:
        - name: invitation_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Invitation declined
  /teams/{id}/invitations:
    post:
      summary: Create team invitation
      operationId: createTeamInvitation
      tags: [Invitations]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                invitee_user_id: { type: string, format: uuid }
              required: [invitee_user_id]
      responses:
        "201":
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamInvitation"

    get:
      summary: List team invitations
      operationId: listTeamInvitations
      tags: [Invitations]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: List of team invitations
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/TeamInvitation"
                required: [items]
  /teams/{id}/invitations/{invitation_id}:
    delete:
      summary: Delete team invitation
      operationId: deleteTeamInvitation
      tags: [Invitations]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: invitation_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Invitation deleted
  /invitations/teams/{invitation_id}/accept:
    post:
      summary: Accept team invitation
      operationId: acceptTeamInvitation
      tags: [Invitations]
      parameters:
        - name: invitation_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Invitation accepted
  /invitations/teams/{invitation_id}/decline:
    post:
      summary: Decline team invitation
      operationId: declineTeamInvitation
      tags: [Invitations]
      parameters:
        - name: invitation_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Invitation declined
...